split in smaller chunks
